+++ AddOns/Cell/RaidFrames/UnitButton.lua
@@ -218,7 +218,7 @@ local function HandleIndicators(b)
             B:UpdateHealthText(b)
         end
         -- update color
-        if t["color"] then
+        if t["color"] and t["indicatorName"] ~= "nameText" then
             indicator:SetColor(unpack(t["color"]))
         end
         -- update colors
@@ -240,6 +240,10 @@ local function HandleIndicators(b)
         -- update duration
         if type(t["showDuration"]) == "boolean" or type(t["showDuration"]) == "number" then
             indicator:ShowDuration(t["showDuration"])
+        end
+        -- update animation
+        if type(t["showAnimation"]) == "boolean" then
+            indicator:ShowAnimation(t["showAnimation"])
         end
         -- update stack
         if type(t["showStack"]) == "boolean" then
@@ -278,6 +282,10 @@ local function HandleIndicators(b)
         -- tooltip
         if type(t["showTooltip"]) == "boolean" then
             indicator:ShowTooltip(t["showTooltip"])
+        end
+        -- blacklist shortcut
+        if type(t["enableBlacklistShortcut"]) == "boolean" then
+            indicator:EnableBlacklistShortcut(t["enableBlacklistShortcut"])
         end
         -- speed
         if t["speed"] then
@@ -510,18 +518,20 @@ local function UpdateIndicators(layout, indicatorName, setting, value, value2)
                 B:UpdateHealthText(b)
             end, true)
         elseif setting == "color" then
-            F:IterateAllUnitButtons(function(b)
-                local indicator = b.indicators[indicatorName]
-                indicator:SetColor(unpack(value))
-            end, true)
+            if indicatorName == "nameText" then
+                F:IterateAllUnitButtons(function(b)
+                    UnitButton_UpdateNameColor(b)
+                end, true)
+            else
+                F:IterateAllUnitButtons(function(b)
+                    local indicator = b.indicators[indicatorName]
+                    indicator:SetColor(unpack(value))
+                end, true)
+            end
         elseif setting == "customColors" then --! NOTE: 其他的colors不调用widget.func，不发出通知，因为这些指示器都使用OnUpdate更新颜色。
             F:IterateAllUnitButtons(function(b)
                 local indicator = b.indicators[indicatorName]
                 indicator:SetColors(value)
-            end, true)
-        elseif setting == "nameColor" then
-            F:IterateAllUnitButtons(function(b)
-                UnitButton_UpdateNameColor(b)
             end, true)
         elseif setting == "vehicleNamePosition" then
             F:IterateAllUnitButtons(function(b)
@@ -625,6 +635,11 @@ local function UpdateIndicators(layout, indicatorName, setting, value, value2)
                     b.indicators[indicatorName]:ShowStack(value2)
                     UnitButton_UpdateAuras(b)
                 end, true)
+            elseif value == "showAnimation" then
+                F:IterateAllUnitButtons(function(b)
+                    b.indicators[indicatorName]:ShowAnimation(value2)
+                    UnitButton_UpdateAuras(b)
+                end, true)
             elseif value == "trackByName" then
                 F:IterateAllUnitButtons(function(b)
                     UnitButton_UpdateAuras(b)
@@ -632,6 +647,10 @@ local function UpdateIndicators(layout, indicatorName, setting, value, value2)
             elseif value == "showTooltip" then
                 F:IterateAllUnitButtons(function(b)
                     b.indicators[indicatorName]:ShowTooltip(value2)
+                end, true)
+            elseif value == "enableBlacklistShortcut" then
+                F:IterateAllUnitButtons(function(b)
+                    b.indicators[indicatorName]:EnableBlacklistShortcut(value2)
                 end, true)
             elseif value == "circledStackNums" then
                 F:IterateAllUnitButtons(function(b)
@@ -694,8 +713,12 @@ local function UpdateIndicators(layout, indicatorName, setting, value, value2)
                     indicator:SetTexture(value["texture"])
                 end
                 -- update showDuration
-                if value["showDuration"] then
+                if type(value["showDuration"]) ~= "nil" then
                     indicator:ShowDuration(value["showDuration"])
+                end
+                -- update showAnimation
+                if type(value["showAnimation"]) == "boolean" then
+                    indicator:ShowAnimation(value["showAnimation"])
                 end
                 -- update showStack
                 if value["showStack"] then
@@ -1031,6 +1054,7 @@ local function UnitButton_UpdateDebuffs(self)
                 -- start, duration, debuffType, texture, count
                 self.indicators.debuffs[startIndex]:SetCooldown((auraInfo.expirationTime or 0) - auraInfo.duration, auraInfo.duration, auraInfo.dispelName or "", auraInfo.icon, auraInfo.applications, refreshing, true)
                 self.indicators.debuffs[startIndex].index = self._debuffs_indices[auraInstanceID] -- NOTE: for tooltip
+                self.indicators.debuffs[startIndex].spellId = auraInfo.spellId -- NOTE: for blacklist
                 startIndex = startIndex + 1
             end
         end
@@ -1041,6 +1065,7 @@ local function UnitButton_UpdateDebuffs(self)
                 -- start, duration, debuffType, texture, count
                 self.indicators.debuffs[startIndex]:SetCooldown((auraInfo.expirationTime or 0) - auraInfo.duration, auraInfo.duration, auraInfo.dispelName or "", auraInfo.icon, auraInfo.applications, refreshing)
                 self.indicators.debuffs[startIndex].index = self._debuffs_indices[auraInstanceID] -- NOTE: for tooltip
+                self.indicators.debuffs[startIndex].spellId = auraInfo.spellId -- NOTE: for blacklist
                 startIndex = startIndex + 1
             end
         end
@@ -1050,6 +1075,7 @@ local function UnitButton_UpdateDebuffs(self)
     self.indicators.debuffs:UpdateSize(startIndex - 1)
     for i = startIndex, 10 do
         self.indicators.debuffs[i].index = nil
+        self.indicators.debuffs[i].spellId = nil
     end
 
     -- update dispels
@@ -1718,17 +1744,12 @@ UnitButton_UpdatePowerMax = function(self)
     if not unit then return end
 
     local value = UnitPowerMax(unit)
-    if value > 0 then
-        if barAnimationType == "Smooth" then
-            self.widget.powerBar:SetMinMaxSmoothedValue(0, value)
-        else
-            self.widget.powerBar:SetMinMaxValues(0, value)
-        end
-        self.widget.powerBar:Show()
-        self.widget.powerBarLoss:Show()
+    if value < 0 then value = 0 end
+    
+    if barAnimationType == "Smooth" then
+        self.widget.powerBar:SetMinMaxSmoothedValue(0, value)
     else
-        self.widget.powerBar:Hide()
-        self.widget.powerBarLoss:Hide()
+        self.widget.powerBar:SetMinMaxValues(0, value)
     end
 end
 
@@ -2088,23 +2109,23 @@ UnitButton_UpdateNameColor = function(self)
         elseif UnitIsCharmed(unit) then
             nameText:SetColor(F:GetClassColor(self.state.class))
         else
-            if Cell.vars.currentLayoutTable["indicators"][1]["nameColor"][1] == "class_color" then
+            if Cell.vars.currentLayoutTable["indicators"][1]["color"][1] == "class_color" then
                 nameText:SetColor(F:GetClassColor(self.state.class))
             else
-                nameText:SetColor(unpack(Cell.vars.currentLayoutTable["indicators"][1]["nameColor"][2]))
+                nameText:SetColor(unpack(Cell.vars.currentLayoutTable["indicators"][1]["color"][2]))
             end
         end
     elseif string.find(unit, "pet") then -- pet
-        if Cell.vars.currentLayoutTable["indicators"][1]["nameColor"][1] == "class_color" then
+        if Cell.vars.currentLayoutTable["indicators"][1]["color"][1] == "class_color" then
             nameText:SetColor(0.5, 0.5, 1)
         else
-            nameText:SetColor(unpack(Cell.vars.currentLayoutTable["indicators"][1]["nameColor"][2]))
+            nameText:SetColor(unpack(Cell.vars.currentLayoutTable["indicators"][1]["color"][2]))
         end
     else -- npc
-        if Cell.vars.currentLayoutTable["indicators"][1]["nameColor"][1] == "class_color" then
+        if Cell.vars.currentLayoutTable["indicators"][1]["color"][1] == "class_color" then
             nameText:SetColor(0, 1, 0.2)
         else
-            nameText:SetColor(unpack(Cell.vars.currentLayoutTable["indicators"][1]["nameColor"][2]))
+            nameText:SetColor(unpack(Cell.vars.currentLayoutTable["indicators"][1]["color"][2]))
         end
     end
 end
@@ -2448,10 +2469,13 @@ local function UnitButton_OnEvent(self, event, unit, arg)
     end
 end
 
+local timer
 local function EnterLeaveInstance()
-    C_Timer.After(1, function()
+    if timer then timer:Cancel() timer=nil end
+    timer = C_Timer.NewTimer(1, function()
         F:Debug("|cffff1111*** EnterLeaveInstance:|r UnitButton_UpdateAll")
         F:IterateAllUnitButtons(UnitButton_UpdateAll, true)
+        timer = nil
     end)
 end
 Cell:RegisterCallback("EnterInstance", "UnitButton_EnterInstance", EnterLeaveInstance)