+++ AddOns/Cell/CHANGELOG.md
@@ -1,4 +1,12 @@
-[Full Changelog](https://github.com/enderneko/Cell/compare/r211-release...e2b743e3c200d2198f22daac5842f2fe4d1e27a7)
+[Full Changelog](https://github.com/enderneko/Cell/compare/r214-release...789b4be076133e29d6948e60e1c2fae62bdf2601)
 
-- Fix newly created indicator issue
-- Update texture of overShieldGlow and overAbsorbGlow
+- Implement blacklist shortcut for Debuffs
+- Add Show Animation option for some indicators
+- Add custom color option for Tank Active Mitigation indicator
+- Update party role order option (#102)
+- Set alwaysUpdateDebuffs to true
+- Fix specFilter of QuickAssist
+- Fix layout preview mover
+- Adjust transparency of Status Text
+- Disable cooldown text from other addons
+- Update RaidDebuffs_WotLK